+++ bookstore/frontend/src/pages/ResetPassword.tsx
@@ -3,10 +3,8 @@ import styled from "styled-components";
 import Title from "../components/common/Title";
 import InputText from "../components/common/InputText";
 import Button from "../components/common/Button";
-import { Link, useNavigate } from "react-router-dom";
-import { resetRequest } from "../api/auth.api";
-import { useAlert } from "../hooks/useAlert";
-import { useState } from "react";
+import { Link } from "react-router-dom";
+import { useAuth } from "@/hooks/useAuth";
 
 export interface ResetPasswordProps {
   email: string;
@@ -14,9 +12,7 @@ export interface ResetPasswordProps {
 }
 
 const ResetPassword = () => {
-  const navigate = useNavigate();
-  const { showAlert } = useAlert();
-  const [resetRequested, setResetRequested] = useState(false);
+  const { userResetPassword, userResetRequest, resetRequested } = useAuth();
 
   const {
     register,
@@ -26,14 +22,9 @@ const ResetPassword = () => {
 
   const onSubmit = (data: ResetPasswordProps) => {
     if (resetRequested) {
-      resetRequest(data).then(() => {
-        showAlert("비밀번호 초기화되었습니다.");
-        navigate("/login");
-      });
+      userResetPassword(data);
     } else {
-      resetRequest(data).then(() => {
-        setResetRequested(true);
-      });
+      userResetRequest(data);
     }
   };
